requirements.txt
absl_py==1.0.0
absl_py==2.0.0
cvxopt==1.3.2
gymnasium==0.29.1
numpy==1.21.5
numpy==1.24.4
numpy==1.24.3
pettingzoo==1.24.1
pygambit==16.1.0a4
python-constraint==1.4.0
Shimmy==1.2.1
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
tianshou==0.5.1
torch==2.0.1
